Added user stories as a set of goals for user interaction in the terminal interface.
Reorganized main in controller.cpp so that the vector holding the initial
values (from input file or stdin) has only the needed lifetime, instead
of hanging around for the entire execution.

Added candidate handling, but it segfaults.  Setting it aside for now
to go on and revise it to use set or bitset.

dcc 11-Feb-2014